Added sleep at sender to reduce the loss of packets

+++ shared.h
@@ -53,7 +53,7 @@ extern long noOfPacketsSent[NUM_UDP_CONNECTION];
 extern long uptoPacketSent[NUM_UDP_CONNECTION] ;
 extern long noOfAckSent[NUM_UDP_CONNECTION] ;
 extern long noOfAckRecd[NUM_UDP_CONNECTION] ;
-extern long noOfLossPackets ;
+extern long long noOfLossPackets ;
 extern long udpSendWaitCount[NUM_UDP_CONNECTION];
 
 extern list<tcpMessage > tcpMessageQ ;
